requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
gunicorn>=20.1.0
pymongo
pydantic
firebase_admin
python-multipart==0.0.6
python-dotenv==1.0.0